moving type checking to its own module
