+++ lib/type/types.js
@@ -1,11 +1,6 @@
 const Type = require('./type');
 const TypeCache = require('./type-cache');
-
-const objectTag = '[object Object]';
-const dateTag = '[object Date]';
-const stringTag = '[object String]';
-const numberTag = '[object Number]';
-const booleanTag = '[object Boolean]';
+const JsTypeChecker = require('./js-type-checker');
 
 const InvoiceDetailDesign = require('../../designs/InvoiceDetail.json');
 const OrderDetailDesign = require('../../designs/OrderDetail.json');
@@ -17,7 +12,6 @@ const OrderDetailType = new Type('OrderDetail', OrderDetailDesign);
 const PackingListDetailType = new Type('PackingListDetail', PackingListDetailDesign);
 const ShippingOrderType = new Type('ShippingOrder', ShippingOrderDesign);
 
-
 /**
  * Instance of all known / declared AppX Types.
  */
@@ -27,7 +21,6 @@ const APPX_TYPES = new TypeCache().registerAll([
     PackingListDetailType,
     ShippingOrderType,
 ]);
-
 
 /**
  * isAppXObject - determines if the object in question has the markings of an
@@ -63,7 +56,6 @@ function getTypeDefinition(o) {
     return isAppXObject(o) ? APPX_TYPES.get(getAppXType(o)) : null;
 }
 
-
 /**
  * isRecognizableAppXType - returns true if o isAppXObject and it's type has
  *                          been registered with APPX_TYPES.
@@ -75,56 +67,10 @@ function isRecognizableAppXType(o) {
   return isAppXObject(o) ? APPX_TYPES.hasKey(getAppXType(o)) : false;
 }
 
-function isString(x) {
-    return typeof x == `string` ?
-        true :
-        (x instanceof String ?
-            true :
-            (Object.prototype.toString.apply(x) === stringTag ?
-                true :
-                false));
-}
-
-function isNumber(x) {
-    return typeof x === `number` ?
-        true :
-        (x instanceof Number ?
-            true :
-            (Object.prototype.toString.apply(x) === numberTag ?
-                true :
-                false));
-}
-
-function isBoolean(x) {
-    return typeof x === `boolean` ?
-        true :
-        (x instanceof Boolean ?
-            true :
-            (Object.prototype.toString.apply(x) === booleanTag ?
-                true :
-                false));
-}
-
-function isDate(x) {
-    return x instanceof Date ?
-        true :
-        (Object.prototype.toString.apply(x) === dateTag ?
-            true :
-            false);
-}
-
-function isPrimitive(x) {
-    return isString(x) || isNumber(x) || isBoolean(x);
-}
-
-function isArray(x) {
-    return Array.isArray(x);
-}
-
 exports.getAppXType = getAppXType;
 exports.isAppXObject = isAppXObject;
 exports.getTypeDefinition = getTypeDefinition;
 exports.isRecognizableAppXType = isRecognizableAppXType;
-exports.isPrimitive = isPrimitive;
-exports.isDate = isDate;
-exports.isArray = isArray;
+exports.isPrimitive = JsTypeChecker.isPrimitive;
+exports.isDate = JsTypeChecker.isDate;
+exports.isArray = JsTypeChecker.isArray;
